+++ mlp/api.py
@@ -108,7 +108,7 @@ async def wants_json() -> bool:
     return True
 
 
-async def handle_error(err_code='UNKNOWN_ERROR', err_msg=None, code: int = None, exc: Exception = None, **kwargs):
+async def handle_error(err_code='UNKNOWN_ERROR', err_msg=None, code: int = None, VERSION: str = '', exc: Exception = None, **kwargs):
     tpl = empty_if(kwargs.get('template'), 'api_error.html')
 
     res, extra = DictObject(error_code=err_code), {}
@@ -129,6 +129,6 @@ async def handle_error(err_code='UNKNOWN_ERROR', err_msg=None, code: int = None,
 
 
     tpl = await render_template(
-        tpl, **res, data=json.dumps(error_dict(res.error_code, msg=res.error_msg, extra=extra), indent=4), settings=settings,  NOTIE_MESSAGE="api_error"
+        tpl, **res, data=json.dumps(error_dict(res.error_code, msg=res.error_msg, extra=extra), indent=4), settings=settings, VERSION=VERSION, NOTIE_MESSAGE="api_error"
     )
     return tpl, res.status_code